+++ client/src/hooks/index.ts
@@ -1,3 +1,5 @@
 import useLocalStorage from './useLocalStorage';
+import useUser from './useUser';
+import useAuth from './useAuth';
 
-export default { useLocalStorage };
+export { useLocalStorage, useUser, useAuth };
